refactor: move vectorlogger to lib.rs
